requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
idna==3.2
multidict==5.1.0
numpy==1.21.0
pandas==1.3.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.0
yarl==1.6.3
